// frontend/src/api/models/SubtitleConfigurationUI.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Klippers API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SubtitleConfiguration } from './SubtitleConfiguration';
import {
    SubtitleConfigurationFromJSON,
    SubtitleConfigurationFromJSONTyped,
    SubtitleConfigurationToJSON,
    SubtitleConfigurationToJSONTyped,
} from './SubtitleConfiguration';

/**
 * 
 * @export
 * @interface SubtitleConfigurationUI
 */
export interface SubtitleConfigurationUI {
    /**
     * 
     * @type {SubtitleConfiguration}
     * @memberof SubtitleConfigurationUI
     */
    subtitleConfiguration?: SubtitleConfiguration | null;
    /**
     * Video url.
     * @type {string}
     * @memberof SubtitleConfigurationUI
     */
    videoUrl?: string;
    /**
     * Thumbnail url.
     * @type {string}
     * @memberof SubtitleConfigurationUI
     */
    thumbnailUrl?: string;
}

/**
 * Check if a given object implements the SubtitleConfigurationUI interface.
 */
export function instanceOfSubtitleConfigurationUI(value: object): value is SubtitleConfigurationUI {
    return true;
}

export function SubtitleConfigurationUIFromJSON(json: any): SubtitleConfigurationUI {
    return SubtitleConfigurationUIFromJSONTyped(json, false);
}

export function SubtitleConfigurationUIFromJSONTyped(json: any, ignoreDiscriminator: boolean): SubtitleConfigurationUI {
    if (json == null) {
        return json;
    }
    return {
        
        'subtitleConfiguration': json['subtitle_configuration'] == null ? undefined : SubtitleConfigurationFromJSON(json['subtitle_configuration']),
        'videoUrl': json['video_url'] == null ? undefined : json['video_url'],
        'thumbnailUrl': json['thumbnail_url'] == null ? undefined : json['thumbnail_url'],
    };
}

export function SubtitleConfigurationUIToJSON(json: any): SubtitleConfigurationUI {
    return SubtitleConfigurationUIToJSONTyped(json, false);
}

export function SubtitleConfigurationUIToJSONTyped(value?: SubtitleConfigurationUI | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'subtitle_configuration': SubtitleConfigurationToJSON(value['subtitleConfiguration']),
        'video_url': value['videoUrl'],
        'thumbnail_url': value['thumbnailUrl'],
    };
}